add traits
