+++ Host.php
@@ -1,39 +1,8 @@
 <?php
 
+require_once 'CNNTrait.php';
+
 class Host
 {
-	private string $code;
-	private string $name;
-	private string $notes;
-	
-
-	public function getCode(): string
-	{
-		return $this->code;
-	}
-
-	public function setCode(string $code): void
-	{
-		$this->code = $code;
-	}
-
-	public function getName(): string
-	{
-		return $this->name;
-	}
-
-	public function setName(string $name): void
-	{
-		$this->name = $name;
-	}
-
-    public function getNotes(): string
-	{
-		return $this->notes;
-	}
-
-	public function setNotes(string $notes): void
-	{
-		$this->notes = $notes;
-	}
+	use CnnTrait;
 }
